drought_model/requirements.txt
azure-storage-blob==12.8.1
azure-core==1.14.0
blinker==1.4
brotlipy==0.7.0
#ca-certificates==2020.10.14
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
idna==3.1
importlib-metadata==4.6.4
intel-openmp==2021.3.0
isodate==0.6.0
# libblas==3.9.0 
# libcblas==3.9.0 
# liblapack==3.9.0 
mkl==2021.3.0
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1 
# openssl==1.1.1k
pandas==1.3.2
pip==21.0.1
pycparser==2.20
pyjwt==2.1.0 
pyopenssl==20.0.1 
pysocks==1.7.1
# python==3.7.11
python-dateutil==2.8.2 
python-dotenv==0.19.0 
# python_abi==3.7
pytz==2021.1 
requests==2.26.0 
requests-oauthlib==1.3.0
setuptools==52.0.0
six==1.16.0
# sqlite==3.36.0
tbb==2021.3.0
typing_extensions==3.10.0.0
urllib3==1.26.6 
# vc==14.2
# vs2015_runtime==14.27.29016
wheel==0.37.0
win_inet_pton==1.1.0
wincertstore==0.2
xgboost==1.4.2
zipp==3.5.0
